refactor: Card 내부 클래스에서 맡은 카드들 생성 기능을 CardFactory에 위임
